+++ src/features/auth/Login.tsx
@@ -7,26 +7,16 @@ import useRedirectWhenLogged from "../../hooks/use-redirect-when-logged";
 import { useNavigate } from "react-router-dom";
 import LoginForm from "./login-form/LoginForm";
 import Div100vh from 'react-div-100vh';
-import { div100LoginStyles } from "./styles/LoginStyles";
+import { loginProfileStyles } from "./styles/LoginStyles";
 import "../../styles/index.css";
 
 const Login: FunctionComponent = () => {
   const { loading, onSubmit } = useLogin();
 
-  const mediaQueryStyles = {
-    '@media (minWidth: 600px)': {
-      minHeight: '500px',
-    },
-    '@media (minWidth: 900px)': {
-      minHeight: '600px',
-    },
-  };
+
 
   return (
-    <Div100vh style={{
-      ...div100LoginStyles,
-      mediaQueryStyles,
-    }}>
+    <Div100vh style={loginProfileStyles}>
       
         <FormControl>
           <LoginForm disabled={loading} onSubmit={onSubmit} />
