+++ src/features/auth/styles/LoginStyles.tsx
@@ -1,6 +1,15 @@
 import Background from "../images/background.jpeg";
 
-export const div100LoginStyles = {
+const mediaQueryStyles = {
+    '@media (minWidth: 600px)': {
+      minHeight: '500px',
+    },
+    '@media (minWidth: 900px)': {
+      minHeight: '600px',
+    },
+  };
+
+export const loginProfileStyles = {
     display: 'flex',
     flexDirection: 'column',
     alignItems: 'center',
@@ -10,4 +19,5 @@ export const div100LoginStyles = {
     backgroundImage: `url(${Background})`,
     backgroundSize: 'cover', 
     backgroundPosition: 'center',
+    ...mediaQueryStyles,
 }
